migrate error　修正

+++ database/migrations/2020_08_21_012914_add_user_id_to_folders.php
@@ -14,7 +14,8 @@ class AddUserIdToFolders extends Migration
     public function up()
     {
         Schema::table('folders', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
+            //integerをbigIにすることでmigrate errorを防ぐ
+            $table->bigInteger('user_id')->unsigned();
 
             // 外部キーを設定する
             $table->foreign('user_id')->references('id')->on('users');
